requirements.txt
python-telegram-bot==20.7
requests>=2.25.0
python-dotenv>=0.19.0